[input] drop fallback window configuration for fullscreeninfo dialog
This will drops the fallback configuration for dialog fullscreeninfo which is currently set to fullscreenvideo.
It is done because it's impossible to navigate inside this modal. Instead of using the default key mapping for left/right/up/down the key mappings of the fallback window are applied.
This is an issue if skins want to use buttons to close or control some elements within this dialog e.g. toggle epg now/next for livetv etc.

(cherry picked from commit e7cec989bb7436c60021d9504c9eb9ee1eeefbeb)
Signed-off-by: Matus Kral <matuskral@me.com>