Wait for vsync before submitting to mmal when display sync is disabled
This avoids an issue where video occasionally goes stuttery after a seek, until the next pause/play or seek.
The issue is when display sync is disabled, and framerate of video matches display, and render times are coincident with vsync
you find that depending on timestamp/scheduling jitter, you may or may not get an update each vsync resulting in stuttery video.

Some scheme to force render times to be dependent on vsync is required. We do this by blocking in RenderUpdate unto next vsync.